[MODREP-25] Undo the changes made for this issue (#117)
For reasons discussed in detail in MODREP-25, it's not really possible in a sensible way to have a tenant-interface handler respond by inserting values into mod-settings, due to its reliance on a `modulePermission` that is defined in the UI module `ui-ldp`.

It would be _possible_ to move the permission across, but scheduling simultaneous major-version updates of `ui-ldp` and `mod-reporting` would be unpleasant enough to make the whole thing not worth doing.

Closes MODREP-25 as WONTFIX.