[MODREP-37] Do not check for errors writing HTTP repsonses
When writing an HTTP response fails, do not attempt to inform the
client by writing the HTTP response again. There is no way to do so
(the connection back to the client is not working due to a timeout or
other error, so there is no point trying to write more to it).

Avoids "superfluous response.WriteHeader call" warning.