[MODREP-46] Disable report-URL whitelist in the example config (#133)
After a very helpful discussion with @wafschneider, I have accepted the philosophically wrong but pragmatically right proposition that the example configuration `etc/config.json` is also, for practical purposes, the default configuration that will be used in many circumstances. For that reason, adding a `reportUrlWhitelist` to the example configuration was in some practical sense a breaking change introduced for MODREP-4.

The present fix disables the default configuration's `reportUrlWhitelist` by renaming it to `DISABLED__reportUrlWhitelist` and documents that i'ts on deployments that actively want security to configure it.

(Some day, we may move this part of the configuration (and maybe more of it?) into the mod-settings store, as suggested by Carole Godfrey in the UILDP-171 discussion.)