[MODREP-42] Configurable timeout, and no silent timeout failure (#137)
Postgres query timeout is configurable both as `queryTimeout` in the config file and (overriding this) in the `MOD_REPORTING_QUERY_TIMEOUT` environment variable. Defaults to 60 seconds if neither is specified.

Also, fixes a bug where reports running for more than 30 seconds would result in the connection to the client being silently dropped.